# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AImanju Backend API",
    "version": "0.1.0",
    "description": "当前仓库已实现路由；标记待接入的路由始终返回 503。服务供应商和前端宽松字段仍需联调。"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/api/credit": {
      "get": {
        "tags": [
          "Credits"
        ],
        "summary": "活跃账号积分余额",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                },
                "example": {
                  "code": 200,
                  "message": "ok",
                  "data": {
                    "total_balance": 1000,
                    "use_credit": 1000,
                    "credit_quota": 1000
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "凭证无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不属于当前账号",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本或幂等冲突；画布冲突包含 data.current_version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "依赖不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/credit/bill": {
      "get": {
        "tags": [
          "Credits"
        ],
        "summary": "分页查询积分流水",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                },
                "example": {
                  "code": 200,
                  "message": "ok",
                  "data": {
                    "list": [],
                    "total": 0
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "凭证无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不属于当前账号",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本或幂等冲突；画布冲突包含 data.current_version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "依赖不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ]
      }
    },
...
    "/api/node/models": {
      "get": {
        "tags": [
          "Generation"
        ],
        "summary": "可用模型和模式能力",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                },
                "example": {
                  "code": 200,
                  "message": "ok",
                  "data": {}
                }
              }
            }
          },
          "400": {
            "description": "请求参数无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "凭证无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不属于当前账号",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本或幂等冲突；画布冲突包含 data.current_version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "依赖不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/node/credit": {
      "post": {
        "tags": [
          "Generation"
        ],
        "summary": "生成询价，不扣费",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                },
                "example": {
                  "code": 200,
                  "message": "ok",
                  "data": {
                    "credit": 5,
                    "capability_id": "mock-image:text2image:v1"
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "凭证无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不属于当前账号",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本或幂等冲突；画布冲突包含 data.current_version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "依赖不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerationCreate"
              }
            }
          }
        }
      }
    },
    "/api/task/generation/create": {
      "post": {
        "tags": [
          "Generation"
        ],
        "summary": "单次生成受理，使用 request_id 幂等",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                },
                "example": {
                  "code": 200,
                  "message": "ok",
                  "data": {
                    "task_id": "00000000-0000-4000-8000-000000000000",
                    "request_id": "example",
                    "status": "queued",
                    "progress": 0
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "凭证无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不属于当前账号",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本或幂等冲突；画布冲突包含 data.current_version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "依赖不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerationCreate"
              }
            }
          }
        }
      }
    },
    "/api/task/generation/batch_create": {
      "post": {
        "tags": [
          "Generation"
        ],
        "summary": "批量生成受理，顺序稳定",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                },
                "example": {
                  "code": 200,
                  "message": "ok",
                  "data": {}
                }
              }
            }
          },
          "400": {
            "description": "请求参数无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "凭证无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不属于当前账号",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本或幂等冲突；画布冲突包含 data.current_version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "依赖不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerationBatch"
              }
            }
          }
        }
      }
    },
    "/api/task/generation/progress": {
      "post": {
        "tags": [
          "Generation"
        ],
        "summary": "任务进度",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                },
                "example": {
                  "code": 200,
                  "message": "ok",
                  "data": {}
                }
              }
            }
          },
          "400": {
            "description": "请求参数无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "凭证无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不属于当前账号",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本或幂等冲突；画布冲突包含 data.current_version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "依赖不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskIds"
              }
            }
          }
        }
      }
    },
    "/api/task/generation/cancel": {
      "post": {
        "tags": [
          "Generation"
        ],
        "summary": "取消排队任务（表单编码）",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                },
                "example": {
                  "code": 200,
                  "message": "ok",
                  "data": {}
                }
              }
            }
          },
          "400": {
            "description": "请求参数无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "凭证无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不属于当前账号",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本或幂等冲突；画布冲突包含 data.current_version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "依赖不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/TaskCancel"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Envelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "ok"
          },
          "data": {}
        },
        "required": [
          "code",
          "message",
          "data"
        ],
        "additionalProperties": false
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "nullable": true
          }
        },
        "required": [
          "code",
          "message"
        ],
        "additionalProperties": false
      },
...
      "GenerationCreate": {
        "type": "object",
        "properties": {
          "request_id": {
            "type": "string"
          },
          "drama_id": {
            "type": "integer",
            "minimum": 1
          },
          "canvas_id": {
            "type": "integer",
            "minimum": 1
          },
          "node_id": {
            "type": "integer",
            "minimum": 1
          },
          "node_type": {
            "type": "string"
          },
          "task_type": {
            "type": "string"
          },
          "model_code": {
            "type": "string"
          },
          "capability_id": {
            "type": "string"
          },
          "mode_type": {
            "type": "string"
          },
          "schema_version": {
            "type": "integer",
            "minimum": 1
          },
          "model_revision": {
            "type": "integer",
            "minimum": 1
          },
          "scene": {
            "type": "string"
          },
          "inputs": {
            "type": "object"
          },
          "parameters": {
            "type": "object"
          },
          "features": {
            "type": "object"
          }
        },
        "required": [
          "drama_id",
          "canvas_id",
          "node_id",
          "node_type",
          "task_type",
          "model_code",
          "inputs"
        ],
        "additionalProperties": false
      },
      "GenerationTask": {
        "type": "object",
        "properties": {
          "node_type": {
            "type": "string"
          },
          "task_type": {
            "type": "string"
          },
          "model_code": {
            "type": "string"
          },
          "capability_id": {
            "type": "string"
          },
          "mode_type": {
            "type": "string"
          },
          "inputs": {
            "type": "object"
          },
          "parameters": {
            "type": "object"
          },
          "features": {
            "type": "object"
          }
        },
        "required": [
          "node_type",
          "task_type",
          "model_code",
          "inputs"
        ],
        "additionalProperties": false
      },
      "GenerationBatch": {
        "type": "object",
        "properties": {
          "request_id": {
            "type": "string"
          },
          "drama_id": {
            "type": "integer",
            "minimum": 1
          },
          "node_id": {
            "type": "integer",
            "minimum": 1
          },
          "canvas_id": {
            "type": "integer",
            "minimum": 1
          },
          "tasks": {
            "type": "array",
            "minItems": 1,
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/GenerationTask"
            }
          }
        },
        "required": [
          "drama_id",
          "node_id",
          "tasks"
        ],
        "additionalProperties": false
      },
      "TaskIds": {
        "type": "object",
        "properties": {
          "task_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "maxItems": 100
          }
        },
        "required": [
          "task_ids"
        ],
        "additionalProperties": false
      },
      "TaskCancel": {
        "type": "object",
        "properties": {
          "task_id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "task_id"
        ],
        "additionalProperties": false
      },
...
    }
  }
}